Add desktopapp module for Bisq 2 app Convert desktop to library
