+++ docs/sample-run-configs.md
@@ -1,6 +1,6 @@
 ### IntelliJ IDEA: Application Run Configs
 
-Here are a few IntelliJ IDEA run configurations for running two seeds and different desktop nodes.
+Here are a few IntelliJ IDEA run configurations for running two seeds and different desktopapp nodes.
 
 #### Run Config: `Seed_1` (clearnet + tor + i2p)
 
@@ -47,12 +47,12 @@ Copy the `Seed_1` run configuration, rename it to `Seed_2` and change:
 
 Classpath of module (Alt+O)
 ```
-bisq.desktop.main
+bisq.desktopapp.main
 ```
 
 Main Class (Alt+C)
 ```
-bisq.desktop.Main
+bisq.desktopapp.Main
 ```
 
 Program Arguments (Alt+R)
@@ -97,7 +97,7 @@ Copy the `Alice_clear` run configuration, rename it to `Alice_i2p` and change:
 
 * Create a new IntelliJ IDEA run config of type Gradle
 * Choose a config name (e.g. `[gradle] Alice I2P`)
-* Choose Gradle Project as `bisq2:desktop`
+* Choose Gradle Project as `bisq2:desktopapp`
 * Add VM options as necessary, in the format `-Dprop=value` (e.g. `-Dbisq.application.appName=bisq_Alice_i2p`)
 
 
@@ -130,18 +130,18 @@ For example, to start two local seeds, `bisq2_seed1` and `bisq2_seed2`, reachabl
     -Dbisq.networkServiceConfig.seedAddressByTransportType.clear.1=127.0.0.1:8001
 ```
 
-Start a desktop client with:
+Start a desktopapp client with:
 
 ```
 # Using default settings
-./gradlew desktop:run
+./gradlew desktopapp:run
 ```
 
-To start a custom desktop client connecting only to clearnet:
+To start a custom desktopapp client connecting only to clearnet:
 
 ```
 # Local client on clearnet only
-./gradlew desktop:run \
+./gradlew desktopapp:run \
     -Dbisq.application.appName=bisq_Alice_clear \
     -Dbisq.networkServiceConfig.supportedTransportTypes.0=CLEAR \
     -Dbisq.networkServiceConfig.seedAddressByTransportType.clear.0=127.0.0.1:8000 \
