+++ settings.gradle
@@ -12,6 +12,7 @@ include 'application'
 include 'cli'
 include 'common'
 include 'desktop'
+include 'desktopapp'
 include 'contract'
 include 'grpc'
 include 'i2p'
@@ -23,7 +24,7 @@ include 'oracle'
 include 'persistence'
 include 'presentation'
 include 'protocol'
-include 'satoshisquare'
+include 'satoshisquareapp'
 include 'security'
 include 'seed'
 include 'settings'
